% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_words}
\alias{get_words}
\title{get_words}
\usage{
get_words(data, user, remove, add_T1 = TRUE, add_K1 = TRUE)
}
\arguments{
\item{data}{file}

\item{user}{A path to user dictionary}

\item{remove}{remove}

\item{add_T1}{add T1}

\item{add_K1}{add K1}
}
\value{
A list.
}
\description{
get words
}
\author{
Yuanlong Hu
}
