% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coNet.R
\name{plot_coNet}
\alias{plot_coNet}
\title{plot_coNet}
\usage{
plot_coNet(
  graph,
  freq_node = 2,
  freq_edge = 2,
  w1 = 10,
  w2 = 10,
  remove_single = TRUE
)
}
\arguments{
\item{graph}{graph}

\item{freq_node}{freq_node}

\item{freq_edge}{freq_edge}

\item{w1}{w1}

\item{w2}{w2}

\item{remove_single}{remove_single}
}
\value{
A igraph
}
\description{
plot_coNet
}
\author{
Yuanlong Hu
}
