% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{read_bib}
\usage{
read_bib(file, source = "CNKI")
}
\arguments{
\item{file}{file}

\item{source}{one of the "CNKI","WanFang", and "VIP".}
}
\value{
A list.
}
\description{
Read bib
}
\author{
Yuanlong Hu
}
