% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_words.R
\name{recode_words}
\alias{recode_words}
\title{recode_words}
\usage{
recode_words(list, re_data)
}
\arguments{
\item{list}{list}

\item{re_data}{re_data}
}
\value{
A list.
}
\description{
recode_words
}
\author{
Yuanlong Hu
}
