% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_words.R
\name{recode_words}
\alias{recode_words}
\title{recode_words}
\usage{
recode_words(list, rep = NULL, remove = NULL)
}
\arguments{
\item{list}{list}

\item{rep}{rep}

\item{remove}{remove words}
}
\value{
A list.
}
\description{
recode_words
}
\author{
Yuanlong Hu
}
